% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assignment_marks.R
\name{add_Q_total}
\alias{add_Q_total}
\title{Adds question total and also updates assignment total}
\usage{
add_Q_total()
}
\value{
adds Question total
}
\description{
Adds question total and also updates assignment total
}
\examples{
Q_total  <- 0
A_total  <- 0
add_marks(10)
add_Q_total()
}
