% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assignment_marks.R
\name{flush_right}
\alias{flush_right}
\title{helper function to put text right}
\usage{
flush_right(x)
}
\arguments{
\item{x}{}
}
\value{
prints to right
}
\description{
helper function to put text right
}
\examples{
flush_right("bob")
}
