% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_assignment.R
\name{make_assignment}
\alias{make_assignment}
\title{Makes assignment}
\usage{
make_assignment(
  questions,
  header,
  coursename,
  assignment,
  lecturer,
  semester,
  year,
  duedate,
  outfile,
  solution = FALSE
)
}
\arguments{
\item{questions}{vector of questions}

\item{header}{header file}

\item{coursename}{coursename}

\item{assignment}{assignment number}

\item{lecturer}{lecturer}

\item{semester}{semester number}

\item{year}{year}

\item{duedate}{due date}

\item{outfile}{file to spit out Rmd}

\item{solution}{do you want solutions}
}
\value{
writes out Rmd
}
\description{
Makes assignment
}
