% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marks.R
\name{marks}
\alias{marks}
\title{add marks in exam.}
\usage{
marks(x)
}
\arguments{
\item{core}{core mark}

\item{adv}{advanced mark}

\item{ms}{marking scheme}
}
\description{
Uses global variables - see the Rmarkdown exam template for an example
}
\examples{
total  <- 0
marks(3)
total
}
