% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_assignment.R
\name{parse_question}
\alias{parse_question}
\title{Takes question and converts it to markdown}
\usage{
parse_question(question, outfile = NULL, solution = FALSE)
}
\arguments{
\item{question}{question Rmd file}

\item{outfile}{file to put markdown}

\item{solution}{boolean of whether to include solutions or not}
}
\value{
nothing, cats out markdown
}
\description{
Takes question and converts it to markdown
}
\examples{
parse_question("inst/rmarkdown/templates/assignment/skeleton/questions/linear_regression.Rmd")
}
