% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/question_template.R
\name{question_template}
\alias{question_template}
\title{Function to give question template}
\usage{
question_template()
}
\value{
question template
}
\description{
Function to give question template
}
