% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/question_total.R
\name{question_total}
\alias{question_total}
\title{Adds question total to exam}
\usage{
question_total()
}
\description{
Adds question total to exam
}
\examples{
total  <- 0
exam_total  <- 0
marks(3)
marks(4)
question_total()
exam_total
}
