% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standard_exam.R
\name{standard_exam}
\alias{standard_exam}
\title{Standard School of Mathematical Science format.}
\usage{
standard_exam(...)
}
\arguments{
\item{...}{Arguments to \code{rmarkdown::pdf_document}}
}
\value{
R Markdown output format to pass to \code{\link[rmarkdown:render]{render}}
}
\description{
Standard School of Mathematical Science format.
}
