% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assignment_marks.R
\name{add_marks}
\alias{add_marks}
\title{Adds marks to part of question.}
\usage{
add_marks(x)
}
\arguments{
\item{x}{number of marks}
}
\value{
add marks
}
\description{
Adds marks to part of question.
}
\examples{
Q_total  <- 0
add_marks(1)
add_marks(10)
}
